requirements.txt
numpy
torch
gym
gensim
matplotlib
faiss-cpu
scikit-learn
datasets
pandas